+++ clown_sort/util/string_helper.py
@@ -10,3 +10,8 @@ def is_empty(text: str) -> bool:
 
 def spaces_to_underscores(text: str) -> str:
     return text.replace(' ', '_')
+
+
+def exception_str(e: Exception) -> str:
+    """A string with the type and message."""
+    return f"{type(e).__name__}: {e}"